backend/user/urls.py
from django.urls import path, include
from rest_framework import routers
from .views import UserViewSet, login, logout

router = routers.DefaultRouter()
router.register('users', UserViewSet, basename='user')

urlpatterns = [
    path('', include(router.urls)),
    path('login/', login, name='login'),
    path('logout/', logout, name='logout'),
]